backend/requirements.txt
fastapi
uvicorn
python-multipart
moviepy
google-generativeai
python-dotenv
